+++ flake.nix
@@ -43,7 +43,7 @@
 
         yadgar-pkg = python.pkgs.buildPythonApplication {
           pname = "yadgar";
-          version = "5.88.1";
+          version = "5.88.2";
           format = "pyproject";
 
           src = ./.;
@@ -175,7 +175,7 @@
 
             coreVersion = lib.mkOption {
               type = lib.types.str;
-              default = "5.88.1";
+              default = "5.88.2";
               description = "Container image tag for the yadgar core service.";
             };